Refactorying NLP package to Theresa client
