+++ .gitignore
@@ -24,7 +24,7 @@ yarn-error.log*
 packages/nexusgraph-app/.yarn
 packages/nexusgraph-editor/.yarn
 packages/nexusgraph-graph/.yarn
-packages/nexusgraph-nlp/.yarn
+packages/nexusgraph-theresa/.yarn
 
 docs/.yarn/
