+++ packages/nexusgraph-app/src/component/sidebar/FeatureMenu.tsx
@@ -2,7 +2,7 @@
 import { useEffect, useState } from "react";
 import { useDispatch } from "react-redux";
 import { t } from "../../../../nexusgraph-i18n";
-import { NaturalLanguageProcessor } from "../../../../nexusgraph-nlp";
+import { NaturalLanguageProcessor } from "../../../../nexusgraph-theresa";
 import { GraphState, INITIAL_GRAPH_NAME, updateGraphData } from "../../../../nexusgraph-redux";
 import { container, TYPES } from "../../../inversify.config";
 import { FeatureButton, IntelligentAITextarea } from "./styled";
